+++ design-patterns/src/js/services/getService.js
@@ -1,11 +1,11 @@
-export class GetService {
+export default class GetService{
     async getSources() {
-        const response = await fetch(`${URL}/sources?apiKey=${API_KEY}`);
+        const response = await fetch(`${this.URL}/sources?apiKey=${this.API_KEY}`);
         return await response.json();
     };
 
     async getArticle(category, pageSize = 10) {
-        const url = `${URL}/top-headlines?category=${category}&pageSize=${pageSize}&apiKey=${API_KEY}`;
+        const url = `${this.URL}/top-headlines?category=${category}&pageSize=${pageSize}&apiKey=${this.API_KEY}`;
         const response = await fetch(url);
         return await response.json();
     };
@@ -14,7 +14,7 @@ export class GetService {
         const params = [];
         params.push(!(+_source === 0) ? 'sources=' + _source : 'category=general');
         params.push('pageSize=' + pageSize);
-        const url = `${URL}/top-headlines?${params.join('&')}&apiKey=${API_KEY}`;
+        const url = `${this.URL}/top-headlines?${params.join('&')}&apiKey=${this.API_KEY}`;
         const response = await fetch(url);
         return await response.json();
     };
